Remove the jpeg file with the OCWM accessibility
We are removing the jpeg slide with the OCWM meeting accessibility to keep the text directly in the OCWM issue. The text will be automatically added as part of the OCWM Issue template

We are creating a separated md file with the text to better discuss and manage future changes.

The issue template and the md with the accessibility are different text and should be maintained in sync for future changes. Alternatively we can modify the github action that create the OCWM issue to automatically inject the content of the accessibility from the md file.